+++ front/src/views/Kafetegia/components/AlergenoLegenda/AlergenoLegenda.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/tools/mediaQueries.scss';
+@import '../../../../styles/tools/mediaQueries.scss';
 
 .wrapper {
   width: 100%;
