+++ front/src/views/Kafetegia/components/AlergenoLegenda/index.ts
